+++ CHANGELOG.md
@@ -2,6 +2,58 @@
 
 <!-- version list -->
 
+## v1.3.0 (2025-11-01)
+
+### Documentation
+
+- Update README to include link to MusicBrainz Picard and improve clarity
+  ([`b0489a7`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/b0489a7561a02f0e129c4ba392621376c3d0d83c))
+
+### Features
+
+- Add DetermineShelfAction to determine shelf from storage location
+  ([`2e14dd5`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/2e14dd56a5a9cc6af42ccc0af6ef136da10b6a9e))
+
+
+## Unreleased
+
+### Bug Fixes
+
+- No shelf if musicbrainz_albumid is not set
+  ([`f6d6193`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/f6d619379acafa52e0f5928d7ab6201209ef5251))
+
+### Documentation
+
+- Enhance plugin description and add license file
+  ([`9211574`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/9211574b5bff2a745605eff7f48201ba3525cb79))
+
+
+## Unreleased
+
+### Bug Fixes
+
+- No shelf if musicbrainz_albumid is not set
+  ([`f6d6193`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/f6d619379acafa52e0f5928d7ab6201209ef5251))
+
+### Documentation
+
+- Enhance plugin description and add license file
+  ([`9211574`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/9211574b5bff2a745605eff7f48201ba3525cb79))
+
+
+## v1.2.2 (2025-10-31)
+
+### Bug Fixes
+
+- No shelf if musicbrainz_albumid is not set
+  ([`f6d6193`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/f6d619379acafa52e0f5928d7ab6201209ef5251))
+
+### Documentation
+
+- Enhance plugin description and add license file
+  ([`9211574`](https://github.com/nrth3rnlb/picard-plugin-shelves/commit/9211574b5bff2a745605eff7f48201ba3525cb79))
+
+
 ## v1.2.1 (2025-10-30)
 
 ### Bug Fixes